School-PHP1
The project is the realization of a one-page site that was implemented using procedural PHP and with MySQL database. Registration and login were realized with regular expressions on the server side. Login can be realized through two roles, user and administrator. The user role has access to the gallery that is realized with pagination. The user is able to send an email to the administrator. The admin role has access to the admin panel and through adding, modifying and deleting content administrator role has the ability to do administration of pages like gallery, survey and users. The menu is dynamically printed from the database depending on the type of role. The survey and the administration of the survey were realized with JavaScript and Ajax.
Technologies – HTML, CSS,PHP, Ajax, JavaScript, JQuery – lightbox, PHPMailer